* commit '54b9676ea6efb7c2ca66458b1fe3b345b57d2d33': (65 commits)
  __assigned_ordinal fields are longs now
  Added primary keys
  Removing LZ4VMS*Stream, replacing with LZ4Block*Stream
  Also send a gutenberg announcement on the vms-<vip>_nostreams namespaces
  Updated dependency lock file to pick up Gutenberg changes
  Fixed gutenberg/filestore dual publish logic
  Need a separate nostreams HollowProducer.Publisher
  Updated dependency lock files
  Publish via Gutenberg, but add metadata for FileStore
  set PackageMomentData default offsets to -1
  Added missing named list VALID_ED_VIDEOS
  Fix issue with DataSlicer to reflect changes in hollow schema
  cleanup legacy merchstill ordering
  change start/end offsets from seconds to millis
  update vms-hollow-generated-notemplate in vms server project too
  Using new vms-generated-impl (special release) in publish-workflow
  remove commented code
  Remove ArtworkCdn type and use cdnId in ArtworkDerivative
  Added ArtworkSourceString to VMSTransformerHashCodeFinder
  Use DownloadableId reference in FileEncodingData
  ...

+++ business-logic/src/main/java/com/netflix/vms/transformer/hollowoutput/BaseDownloadable.java
@@ -53,5 +53,5 @@ public class BaseDownloadable implements Cloneable {
     }
 
     @SuppressWarnings("unused")
-    private int __assigned_ordinal = -1;
+    private long __assigned_ordinal = -1;
 }
