+++ business-logic/src/main/java/com/netflix/vms/transformer/hollowoutput/VideoAwardFestival.java
@@ -43,5 +43,5 @@ public class VideoAwardFestival implements Cloneable {
     }
 
     @SuppressWarnings("unused")
-    private int __assigned_ordinal = -1;
+    private long __assigned_ordinal = -1;
 }
